구조 분할할
